Listar y borrar usuarios- List of users and delete users
Added the logic to delete user (logical delete) and the test to list users and delete an user. Also fixed some anotations of the brach thatr i use to copy this one

+++ demo/src/main/java/com/project/demo/logic/entity/country/Country.java
@@ -18,7 +18,7 @@ public class Country {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long country_id;
+    private Integer country_id;
 
     @Column(nullable = true)
     private String country_name;
@@ -31,13 +31,14 @@ public class Country {
 
 
     // Constructors
-    public Country() {}
+    public Country() {
+    }
 
-    public Long getCountry_id() {
+    public Integer getCountry_id() {
         return country_id;
     }
 
-    public void setCountry_id(Long country_id) {
+    public void setCountry_id(Integer country_id) {
         this.country_id = country_id;
     }
 
@@ -65,11 +66,7 @@ public class Country {
         this.operational = operational;
     }
 
+    ;
 
-
-//    public Country setRole(Country country) {
-//        this.role = role;
-//
-//        return this;
-//    }
 }
+
